Add files via upload
Added an option to change between 4 different Mapbox map styles (dark, satellite, satellite with streets and dark navigation). On the map there is now a small preview of the next style and by clicking that the map style changes. It also sets the selection to the cookies, so the setting will be remembered.