Add files via upload
Changed button positions a bit on the map.
Added also a COVR button to the map (bottom left) which will toggle receiver coverage lines on and off. This differes slightly to the receivers / antenna circular distance / direction graph, because that is on a flat plane base and the map coverage is in spheroid base.