Add files via upload
Since there was no answer from the OpenSky Network (question if this is ok to be done sent on September), I'll take that as usage approved. Please note that any kind of commercial usage of OpenSky Network data cannot be done. That is only for personal and research purposes.
So added OpenSky Network call to fetch ALL the other aircraft than the dump1090-fa received and a button called SOSN, which will toggle showing those aircraft on the map. The data fetch is done every 15 seconds in order to throttle the API calls to OSN.