Add files via upload
Bugfix: if OpenSky Network icao has been found from the list of primary receiver icaos it was OSN position was not shown. However, the icao didn't necessarily have position information.
Now if the OSN icao is found in primary, but it does'nt have position information, the OSN marker is not deleted, but colored a bit more green. This will indicate that the aircraft could possibly either enter soon the receiver range or is exiting the range.